+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 We use [semantic versioning][semver]
 
 # Next Release
+- [fix] Delete empty coverage directories
 
 # 15.4.0
 - [fix] `git.properties` commit was not used for upload